Traffic_light moved to cross_road private class
